Customized font and interactionText
Added a personalized font (not perfect but good enough for now). In the future use the same structure in aseprite. Consider also finding a way to deal with new line spaces being too big. Added as well a new parameter to LoaderParams to make each sign display personalized interactionText from the XML file